requirements.txt
requests
pyyaml
tqdm
colorama